Add products' index, new, and edit views and logic

+++ db/migrate/20201203163658_create_products.rb
@@ -2,9 +2,9 @@ class CreateProducts < ActiveRecord::Migration[6.0]
   def change
     create_table :products do |t|
       t.string :name
-      t.decimal :price, precision: 8, scale: 2
+      t.float :price, precision: 8, scale: 2
       t.integer :quantity
-      t.integer :upc
+      t.bigint :upc
 
       t.timestamps
     end
